+++ apps/spacerage/src/scenes/gameplay/GameplayScene.tsx
@@ -7,10 +7,10 @@ import { Stage } from "../../configuration"
 import { Nebula } from "../menu/vfx/Nebula"
 import { Asteroids } from "./Asteroids"
 import { Bullets } from "./Bullets"
-import { Debris } from "./Debris"
+import { Debris } from "./vfx/Debris"
 import { FollowCamera } from "./FollowCamera"
 import { Player } from "./Player"
-import { Sparks } from "./Sparks"
+import { Sparks } from "./vfx/Sparks"
 import { AgeSystem } from "./systems/AgeSystem"
 import { BulletSystem } from "./systems/BulletSystem"
 import { DestroyAfterSystem } from "./systems/DestroyAfterSystem"
