+++ apps/spacerage/src/scenes/gameplay/actions.tsx
@@ -1,6 +1,6 @@
 import { Color, Quaternion, Vector3 } from "three"
-import { DebrisEmitter } from "./Debris"
-import { SparksEmitter } from "./Sparks"
+import { DebrisEmitter } from "./vfx/Debris"
+import { SparksEmitter } from "./vfx/Sparks"
 import { ECS } from "./state"
 
 export const spawnAsteroid = (position: Vector3, scale: number = 1) => {
